Rename `server` folder to `start`
