+++ packages/cli/src/commands/start/watchMode.ts
